Started recording version changes + bugs in files. Moved all file IO functions to their own files. Added temporary functions for rewriting Artist related data structures. Fixed bug with disconnect handling. Fixed an issue in handlePlayerTrackSubmission where it wasnt clearing the correct Timeout. Fixed an issue that doesnt update the artists common to the same song. Fixed an issue that let users use the same track multiple times. Fixed an issue where the timer would stop earlier than it should (should stop after all the track invalid checks are done) Fixed an issue where if a client disconnected from the server while in queue, the server would not update the queue properly. Modified internal file structure to be more compact. Moved game related data to it's own file gameData.js
TODO and Bugs:
Need to move case/switch functions and events to their own functions.
Need to change game and websocket events into numbers to save space in packets.
Theres an issue where when a player disconnect, the other player is not sent a win message.
Change the timeline so it's less confusing.
Move the files for connection related information to it's own file.